Per #2547, rename nccf files as nc_cf for consistency throughout
