+++ src/libcode/vx_data2d_factory/data2d_factory.cc
@@ -22,7 +22,7 @@ using namespace std;
 #include "data2d_grib.h"
 #include "data2d_nc_met.h"
 #include "data2d_nc_pinterp.h"
-#include "data2d_nccf.h"
+#include "data2d_nc_cf.h"
 #ifdef WITH_UGRID
 #include "data2d_ugrid.h"
 #endif
