+++ src/libcode/vx_data2d_nc_cf/Makefile.am
@@ -12,9 +12,9 @@ include ${top_srcdir}/Make-include
 
 noinst_LIBRARIES = libvx_data2d_nc_cf.a
 libvx_data2d_nc_cf_a_SOURCES = \
-              nccf_file.cc nccf_file.h \
-              var_info_nccf.cc var_info_nccf.h \
-              data2d_nccf.cc data2d_nccf.h \
+              nc_cf_file.cc nc_cf_file.h \
+              var_info_nc_cf.cc var_info_nc_cf.h \
+              data2d_nc_cf.cc data2d_nc_cf.h \
               vx_data2d_nc_cf.h
 libvx_data2d_nc_cf_a_CPPFLAGS = ${MET_CPPFLAGS}
