+++ src/libcode/vx_data2d_nc_cf/nc_cf_file.cc
@@ -29,7 +29,7 @@ using namespace netCDF;
 #include "vx_cal.h"
 #include "vx_log.h"
 
-#include "nccf_file.h"
+#include "nc_cf_file.h"
 
 
 ////////////////////////////////////////////////////////////////////////
