+++ src/libcode/vx_data2d_nc_cf/nc_cf_file.h
@@ -14,8 +14,8 @@
 ////////////////////////////////////////////////////////////////////////
 
 
-#ifndef  __NCCF_FILE_H__
-#define  __NCCF_FILE_H__
+#ifndef  __NC_CF_FILE_H__
+#define  __NC_CF_FILE_H__
 
 
 ////////////////////////////////////////////////////////////////////////
@@ -194,7 +194,7 @@ class NcCfFile {
 ////////////////////////////////////////////////////////////////////////
 
 
-#endif   /*  __NCCF_FILE_H__  */
+#endif   /*  __NC_CF_FILE_H__  */
 
 
 ////////////////////////////////////////////////////////////////////////
