+++ src/libcode/vx_data2d_nc_cf/vx_data2d_nc_cf.h
@@ -21,9 +21,9 @@
 ////////////////////////////////////////////////////////////////////////
 
 
-#include "data2d_nccf.h"
-#include "var_info_nccf.h"
-#include "nccf_file.h"
+#include "data2d_nc_cf.h"
+#include "var_info_nc_cf.h"
+#include "nc_cf_file.h"
 
 
 ////////////////////////////////////////////////////////////////////////
